Upgrade GSR adapter from v2 to v3 (#2385)
* Upgrade GSR adapter from v2 to v3

* Bump changeset version change to "major"
