+++ .changeset/weak-plums-ring.md
@@ -1,0 +1,5 @@
+---
+'@chainlink/gsr-adapter': major
+---
+
+Upgraded EA framework from v2 to v3
