+++ packages/core/legos/src/sources.ts
@@ -84,7 +84,6 @@ import * as google_bigquery from '@chainlink/google-bigquery-adapter'
 import * as gramchain from '@chainlink/gramchain-adapter'
 import * as graphql from '@chainlink/graphql-adapter'
 import * as gsr from '@chainlink/gsr-adapter'
-import * as gsr_test from '@chainlink/gsr-test-adapter'
 import * as iex_adapter from '@chainlink/iex-cloud-adapter'
 import * as intrinio from '@chainlink/intrinio-adapter'
 import * as ipfs from '@chainlink/ipfs-adapter'
@@ -234,7 +233,6 @@ export default {
   gramchain,
   graphql,
   gsr,
-  gsr_test,
   iex_adapter,
   intrinio,
   ipfs,
