+++ packages/core/legos/tsconfig.json
@@ -269,9 +269,6 @@
       "path": "../../sources/gsr"
     },
     {
-      "path": "../../sources/gsr-test"
-    },
-    {
       "path": "../../sources/iex-cloud"
     },
     {
