--- packages/sources/gsr-test/CHANGELOG.md
@@ -1,32 +1,0 @@
-# @chainlink/gsr-test-adapter
-
-## 1.0.4
-
-### Patch Changes
-
-- 22542130d: Add "crypto" endpoint alias
-
-## 1.0.3
-
-### Patch Changes
-
-- f850f4831: Bump framework version
-- 917400c7b: Bump framework version
-
-## 1.0.2
-
-### Patch Changes
-
-- b0ed53787: Bump framework version
-
-## 1.0.1
-
-### Patch Changes
-
-- 07781d25d: Bumped framework version
-
-## 1.0.0
-
-### Major Changes
-
-- ee928416a: Added GSR-test EA
