--- packages/sources/gsr-test/src/endpoint/index.ts
@@ -1,1 +1,0 @@
-export { endpoint as price } from './price'
