--- packages/sources/gsr-test/test-payload.json
@@ -1,6 +1,0 @@
-{
- "requests": [{
-  "from": "ETH",
-  "to": "USD"
- }]
-}
