--- packages/sources/gsr/schemas/env.json
@@ -1,36 +1,0 @@
-{
-  "$id": "https://external-adapters.chainlinklabs.com/schemas/gsr-adapter.json",
-  "title": "GSR Adapter",
-  "description": "This adapter only supports WS connections. Make sure WS is enabled in your configuration in order to run this adapter.",
-  "required": ["WS_USER_ID", "WS_PUBLIC_KEY", "WS_PRIVATE_KEY"],
-  "type": "object",
-  "properties": {
-    "API_ENDPOINT": {
-      "type": "string",
-      "description": "The endpoint to fetch the GSR access token from (required to establish a websocket connection)",
-      "default": "https://oracle.prod.gsr.io/v1"
-    },
-    "WS_API_ENDPOINT": {
-      "type": "string",
-      "description": "The websocket endpoint to pull data from",
-      "default": "wss://oracle.prod.gsr.io/oracle"
-    },
-    "WS_USER_ID": {
-      "type": "string",
-      "description": "The user ID used to authenticate"
-    },
-    "WS_PUBLIC_KEY": {
-      "type": "string",
-      "description": "The public key used to authenticate"
-    },
-    "WS_PRIVATE_KEY": {
-      "type": "string",
-      "description": "The private key used to authenticate"
-    }
-  },
-  "allOf": [
-    {
-      "$ref": "https://external-adapters.chainlinklabs.com/schemas/ea-bootstrap.json"
-    }
-  ]
-}
