--- packages/sources/gsr/src/config/envDefaultOverrides.ts
@@ -1,5 +1,0 @@
-import type { EnvDefaultOverrides } from '@chainlink/ea-bootstrap'
-
-export const envDefaultOverrides: EnvDefaultOverrides = {
-  WS_ENABLED: 'true',
-}
