--- packages/sources/gsr/src/config/includes.json
@@ -1,1 +1,0 @@
-{}
