+++ packages/tsconfig.json
@@ -381,9 +381,6 @@
       "path": "./sources/gsr"
     },
     {
-      "path": "./sources/gsr-test"
-    },
-    {
       "path": "./sources/iex-cloud"
     },
     {
